Create entity-tile domain dir in components and move EntityTile into it
