+++ context/app/static/js/components/entity-tile/EntityTile/style.js
